Migrated doi page to app router
