+++ package.json
@@ -82,8 +82,8 @@
     "@types/react": "^18.2.48",
     "@types/react-bootstrap": "0.32.32",
     "@types/react-dom": "^18.2.18",
-    "@typescript-eslint/eslint-plugin": "^4.28.0",
-    "@typescript-eslint/parser": "^4.28.0",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
     "babel-core": "^6.26.3",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^26.0.1",
@@ -103,7 +103,7 @@
     "ts-jest": "^26.1.0",
     "ts-loader": "^7.0.5",
     "tsify": "^4.0.2",
-    "typescript": "^4.3.2"
+    "typescript": "^5.3.3"
   },
   "resolutions": {
     "**/@fortawesome/fontawesome-common-types": "^0.3.0"
