+++ src/components/DownloadReports/DownloadReports.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Row, Col } from 'react-bootstrap'
+import { Row, Col } from 'src/components/Layout'
 import HelpIcon from '../HelpIcon/HelpIcon'
