+++ src/components/HorizontalStackedBarChart/HorizontalStackedBarChart.module.scss
@@ -1,7 +1,6 @@
 .chartText {
   max-width: 300px;
 
-  font-family: 'Source Sans Pro';
   font-size: 21px;
   font-weight: 700;
 
@@ -22,7 +21,6 @@
 }
 
 .chartDetailText {
-  font-family: 'Source Sans Pro';
   font-size: 28px !important;
 
   color: #1ABC9C;
