+++ src/components/SankeyGraph/SankeyGraph.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import React, { useEffect, useRef, useState } from 'react'
 import { Vega } from 'react-vega'
