+++ src/components/ShareLinks/ShareLinks.tsx
@@ -1,5 +1,7 @@
+'use client'
+
 import React from 'react'
-import { Row, Col } from 'react-bootstrap'
+import { Row, Col } from 'src/components/Layout'
 import {
   EmailShareButton,
   FacebookShareButton,
