+++ src/data/queries/crossrefFunderQuery.ts
@@ -1,0 +1,9 @@
+import { gql } from '@apollo/client'
+
+export const CROSSREF_FUNDER_GQL = gql`
+  query getOrganizationQuery($crossrefFunderId: ID) {
+    organization(crossrefFunderId: $crossrefFunderId) {
+      id
+    }
+  }
+`
